First complete working version
Added the CribStrategy file, to act as the entrypoint and handle the actual strategy calculations for what to throw out, and the Statistics class, to handle some of the array math.

Changed Hand to Deck, since most of the operations required the whole deck and not just a single hand.

Removed the entrypoint in CribScorer because CribStrategy has a "Score" feature that uses this instead.